Fixes #957

+++ inc/language-manager.php
@@ -87,7 +87,7 @@ class Odm_Language_Manager {
   function print_language_flags_for_post($post){
     
     foreach ($this->supported_languages as $lang_code => $lang):
-      if(qtranxf_isAvailableIn($post->ID, $lang_code)){ // no En content
+      if(function_exists('qtranxf_isAvailableIn') && qtranxf_isAvailableIn($post->ID, $lang_code)){ // no En content
         $path_to_flag = odm_language_manager()->get_path_to_flag_image($lang_code);
         if (!empty($path_to_flag)):
           echo '<img class="lang_flag" alt="' . $lang . '" src="' . $path_to_flag .'"></img>';
